requirements.txt
Flask==3.1.2
gunicorn==23.0.0
google-generativeai==0.8.5